:recycle: api: Use single source for headers
