+++ api/lib/infrastructure/files/sessions-import.js
@@ -1,5 +1,15 @@
+const { headers } = require('../utils/csv/sessions-import');
+
 function getHeaders() {
-  return '"N° de session";"* Nom du site";"* Nom de la salle";"* Date de début";"* Heure de début (heure locale)";"* Surveillant(s)";"Observations (optionnel)";"* Nom de naissance";"* Prénom";"* Date de naissance (format: jj/mm/aaaa)";"* Sexe (M ou F)";"Code Insee";"Code postal";"Nom de la commune";"* Pays";"E-mail du destinataire des résultats (formateur, enseignant…)";"E-mail de convocation";"Identifiant local";"Temps majoré ?"';
+  const fields = _getHeadersAsArray();
+
+  return fields;
+}
+
+function _getHeadersAsArray() {
+  return Object.keys(headers)
+    .reduce((arr, key) => [...arr, `"${headers[key]}"`], [])
+    .join(';');
 }
 
 module.exports = { getHeaders };
