+++ api/lib/infrastructure/serializers/csv/csv-serializer.js
@@ -1,6 +1,7 @@
 const logger = require('../../logger');
 const { FileValidationError } = require('../../../../lib/domain/errors');
 const { convertDateValue } = require('../../utils/date-utils');
+const { headers } = require('../../utils/csv/sessions-import');
 
 function _csvFormulaEscapingPrefix(data) {
   const mayBeInterpretedAsFormula = /^[-@=+]/.test(data);
@@ -23,29 +24,6 @@ function _csvSerializeValue(data) {
 }
 
 function deserializeForSessionsImport(parsedCsvData) {
-  const headers = {
-    address: '* Nom du site',
-    room: '* Nom de la salle',
-    date: '* Date de début',
-    time: '* Heure de début (heure locale)',
-    examiner: '* Surveillant(s)',
-    description: 'Observations (optionnel)',
-    lastName: '* Nom de naissance',
-    firstName: '* Prénom',
-    birthdate: '* Date de naissance (format: jj/mm/aaaa)',
-    sex: '* Sexe (M ou F)',
-    birthINSEECode: 'Code Insee',
-    birthPostalCode: 'Code postal',
-    birthCity: 'Nom de la commune',
-    birthCountry: '* Pays',
-    resultRecipientEmail: 'E-mail du destinataire des résultats (formateur, enseignant…)',
-    email: 'E-mail de convocation',
-    externalId: 'Identifiant local',
-    extraTimePercentage: 'Temps majoré ?',
-    billingMode: 'Tarification part Pix',
-    prepaymentCode: 'Code de prépaiement',
-  };
-
   const sessions = [];
   const csvLineKeys = Object.keys(headers);
